requirements.txt
anyio==4.9.0
boto3==1.37.33
botocore==1.37.33
certifi==2025.1.31
fire==0.7.0
fonttools==4.57.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jmespath==1.0.1
lxml==5.3.2
numpy==2.2.4
opencv-python-headless==4.11.0.86
pdf2docx==0.5.8
prometheus_client==0.21.1
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-telegram-bot==22.0
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
termcolor==3.0.1
typing_extensions==4.13.2
urllib3==2.4.0
